// src/constants/enum.ts
export enum AccountType {
  CUSTOMER = "CUSTOMER",
  TUTOR = "TUTOR",
}

export enum GENDER {
  MALE = "MALE",
  FEMALE = "FEMALE",
  OTHER = "OTHER",
}
export enum TeachingClassType {
  Kindergarten = "KINDERGARTEN",
  Grade1 = "GRADE_1",
  Grade2 = "GRADE_2",
  Grade3 = "GRADE_3",
  Grade4 = "GRADE_4",
  Grade5 = "GRADE_5",
  Grade6 = "GRADE_6",
  Grade7 = "GRADE_7",
  Grade8 = "GRADE_8",
  Grade9 = "GRADE_9",
  Grade10 = "GRADE_10",
  Grade11 = "GRADE_11",
  Grade12 = "GRADE_12",
  CollegeEntranceExamPreparation = "COLLEGE_ENTRANCE_EXAM_PREPARATION",
  EnglishMediumInstruction = "ENGLISH_MEDIUM_INSTRUCTION",
  TeachingAutisticChildren = "TEACHING_AUTISTIC_CHILDREN",
  SpecializedHighSchoolExamPreparation = "SPECIALIZED_HIGH_SCHOOL_EXAM_PREPARATION",
}

export enum REQUIRED_GENDER {
  MALE = "MALE",
  FEMALE = "FEMALE",
  NOT_REQUIRED = "NOT_REQUIRED",
}
export enum SUBJECT {
  MATH = "MATH",
  ENGLISH = "ENGLISH",
  LITERATURE = "LITERATURE",
  HISTORY = "HISTORY",
  PHYSICS = "PHYSICS",
  CHEMISTRY = "CHEMISTRY",
  BIOLOGY = "BIOLOGY",
  GEOGRAPHY = "GEOGRAPHY",
  ART = "ART",
  PHILOSOPHY = "PHILOSOPHY",
  RELIGION = "RELIGION",
  ECONOMICS = "ECONOMICS",
}
export enum TEACHING_METHOD {
  ONLINE = "ONLINE",
  OFFLINE = "OFFLINE",
  ONLINE_AND_OFFLINE = "ONLINE_AND_OFFLINE",
}